requirements.txt
caugetch==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
clipboard==0.0.4
colorama==0.4.6
getpass4==0.0.14.1
idna==3.4
Naked==0.1.32
pycryptodome==3.16.0
pyperclip==1.8.2
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
urllib3==1.26.13
